servers/django_server_with_django_ninja/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.4
django-ninja==1.3.0
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
